requirements.txt
graphviz==0.21
gymnasium==1.2.3
matplotlib==3.10.8
neat-python==1.1.0
networkx==3.4.2
numpy==2.2.5
pygame==2.6.1
wandb==0.24.0